# src/app.ts
import apicache from 'apicache';
import cors from 'cors';
import express, { type NextFunction, type Response } from 'express';
import rateLimit from 'express-rate-limit';
import config from './config';
import * as leetcode from './leetCode';
import { setupSwagger } from './swagger';
import type { FetchUserDataRequest } from './types';

const app = express();
app.set('trust proxy', config.trustProxy);
const cache = apicache.middleware;

const limiter = rateLimit({
  windowMs: 60 * 60 * 1000, // 1 hour
  limit: 120,
  standardHeaders: 'draft-7',
  legacyHeaders: false,
  message: 'Too many request from this IP, try again in 1 hour',
});

app.use(cache('5 minutes'));
app.use(cors()); //enable all CORS request
app.use(limiter); //limit to all API
app.use((req: express.Request, _res: Response, next: NextFunction) => {
  console.log('Requested URL:', req.originalUrl);
  next();
});

// Swagger UI must be mounted before /:username wildcard routes
setupSwagger(app);

app.get('/', (_req, res) => {
  res.json({
    apiOverview:
      'Welcome to the Alfa-Leetcode-API! Alfa-Leetcode-Api is a custom solution born out of the need for a well-documented and detailed LeetCode API. This project is designed to provide developers with endpoints that offer insights into a user"s profile, badges, solved questions, contest details, contest history, submissions, and also daily questions, selected problem, list of problems.',
    apiEndpointsLink:
      'https://github.com/alfaarghya/alfa-leetcode-api?tab=readme-ov-file#endpoints-',
    routes: {
      userDetails: {
        description:
          'Endpoints for retrieving detailed user profile information on Leetcode.',
        Method: 'GET',
        '/:username': 'Get your leetcode profile Details',
        '/:username/profile': 'Get full profile details',
        '/:username/badges': 'Get your badges',
        '/:username/solved': 'Get total number of question you solved',
        '/:username/contest': 'Get your contest details',
        '/:username/contest/history': 'Get all contest history',
        '/:username/submission': 'Get your last 20 submission',
        '/:username/submission?limit=7':
          'Get a specified number of last submissions.',
        '/:username/acSubmission': 'Get your last 20 accepted submission',
        '/:username/acSubmission?limit=7':
          'Get a specified number of last acSubmissions.',
        '/:username/calendar': 'Get your submission calendar',
        '/:username/calendar?year=2025': 'Get your year submission calendar',
        '/:username/skill': 'Get your skill stats',
        '/:username/language': 'Get your language stats',
        '/:username/progress': 'Get your progress stats',
      },
      discussion: {
        description: 'Endpoints for fetching discussion topics and comments.',
        Method: 'GET',
        '/trendingDiscuss?first=20': 'Get top 20 trending discussions',
        '/discussTopic/:topicId': 'Get discussion topic',
        '/discussComments/:topicId': 'Get discussion comments',
      },
      problems: {
        description:
          'Endpoints for fetching problem-related data, including lists, details, and solutions.',
        Method: 'GET',
        singleProblem: {
          '/select?titleSlug=two-sum': 'Get selected Problem',
          '/select/raw?titleSlug=two-sum': 'Get raw selected Problem',
          '/daily': 'Get daily Problem',
          '/daily/raw': 'Get raw daily Problem',
        },
        '/tags': 'Get list of all problem topic tags',
        problemList: {
          '/problems': 'Get list of 20 problems',
          '/problems?limit=50': 'Get list of some problems',
          '/problems?tags=array+math': 'Get list problems on selected topics',
          '/problems?tags=array+math+string&limit=5':
            'Get list some problems on selected topics',
          '/problems?skip=500':
            'Get list after skipping a given amount of problems',
          '/problems?difficulty=EASY':
            'Get list of problems having selected difficulty',
          '/problems?limit=5&skip=100':
            'Get list of size limit after skipping selected amount',
          'problems?tags=array+maths&limit=5&skip=100':
            'Get list of problems with selected tags having size limit after skipping selected amount',
          '/officialSolution?titleSlug=two-sum':
            'Get official solution of selected problem',
        },
      },
    },
  });
});

//get trending Discuss
app.get('/trendingDiscuss', leetcode.trendingCategoryTopics);

//get discuss topic
app.get('/discussTopic/:topicId', leetcode.discussTopic);

//get discuss comments
app.get('/discussComments/:topicId', leetcode.discussComments);

//get the daily leetCode problem
app.get('/daily', leetcode.dailyProblem);
app.get('/daily/raw', leetcode.dailyProblemRaw);

//get the selected question
app.get('/select', leetcode.selectProblem);
app.get('/select/raw', leetcode.selectProblemRaw);

//get official solution
app.get('/officialSolution', leetcode.officialSolution);

//get list of all topic tags
app.get('/tags', leetcode.tags);

//get list of problems
app.get('/problems', leetcode.problems);

//get contests
app.get('/contests', leetcode.allContests);
app.get('/contests/upcoming', leetcode.upcomingContests);

// Construct options object on all user routes.
app.use(
  '/:username*',
  (req: FetchUserDataRequest, _res: Response, next: NextFunction) => {
    req.body = {
      username: req.params.username,
      limit: req.query.limit ? parseInt(req.query.limit as string, 10) : 20,
      year: req.query.year ? parseInt(req.query.year as string, 10) : 0,
    };
    next();
  },
);

//get user profile details
app.get('/:username', leetcode.userData);

//get user badges
app.get('/:username/badges', leetcode.userBadges);

//get solved problem
app.get('/:username/solved', leetcode.solvedProblem);

//get user contest details
app.get('/:username/contest', leetcode.userContest);

//get user contest history
app.get('/:username/contest/history', leetcode.userContestHistory);

//get user submission
app.get('/:username/submission', leetcode.submission);

//get user acSubmission
app.get('/:username/acSubmission', leetcode.acSubmission);

//get user calendar
app.get('/:username/calendar', leetcode.calendar);

//get user skill stats
app.get('/:username/skill/', leetcode.skillStats);

//get user profile
app.get('/:username/profile/', leetcode.userProfile);

//get user language stats
app.get('/:username/language', leetcode.languageStats);

//get user progress
app.get('/:username/progress/', leetcode.progress);

/* ----- Migrated to new routes -> these will be deleted ----- */

app.get('/userProfile/:id', leetcode.userProfile_);
app.get('/dailyQuestion', leetcode.dailyQuestion_);
app.get('/selectQuestion', leetcode.selectProblemRaw);
app.get('/skillStats/:username', leetcode.skillStats_);
app.get(
  '/userProfileUserQuestionProgressV2/:userSlug',
  leetcode.userProfileUserQuestionProgressV2_,
);
app.get('/languageStats', leetcode.languageStats_);
app.get('/userContestRankingInfo/:username', leetcode.userContestRankingInfo_);

export default app;